pyproject.toml
[tool.poetry]
name = "code-meeting-auto-update"
version = "0.1.0"
description = ""
authors = ["as6325400 <as6325400@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
